+++ src/visor/dbutil/dbutil.go
@@ -345,7 +345,7 @@ func Reset(tx *Tx, bktName []byte) error {
 // Itob converts uint64 to bytes
 func Itob(v uint64) []byte {
 	b := make([]byte, 8)
-	binary.BigEndian.PutUint64(b, uint64(v))
+	binary.BigEndian.PutUint64(b, v)
 	return b
 }
